feat: [문장] 사용자 메인 문장 제거
- move reading sentence api to sentence controller

https://inoutedu.atlassian.net/browse/IO-18
